dsp-kernel: Avoid dereferencing user pointer in the kernel
When the user invokes the FASTRPC_IOCTL_INIT_CREATE IOCTL with
argp.file set to an invalid user pointer, it is considered a non-fd
input buffer. Consequently, fastrpc_get_args copies data from
args.file to the header buffer, which is kernel-mapped. Because the
kernel parameter is set, instead of using copy_from_user, a call to
memcpy is performed with the user-controlled address as the source,
which can cause a device crash. Return the -EINVAL error upon
receiving the invalid file descriptor.

Change-Id: I982e3af17a75ae4a16081de958bf5840ad09a7b5
Signed-off-by: Abhishek Singh <quic_abhishes@quicinc.com>