qcacld-3.0: Check pending stats request before requesting new one
Currently due to a race condition between ll_stats and get_station
stats request, the existing check to see if there is any pending
stats request is not sufficient. This is leading to host requesting
multiple stats request and the the callbacks of the stats request
getting overwritten.

To resolve this race condition, add an additional check for pending
stats request, before sending a new one to firmware.

Change-Id: I801f6e2d4459aa45024020fa34ac202e67f33d6d
CRs-Fixed: 3939986