dsp-kernel: Prevent buffer overflow with buffer size consistency
When the user calls the IOCTL cases FASTRPC_IOCTL_MULTIMODE_INVOKE and
FASTRPC_IOCTL_INIT_CREATE on separate threads with changing values of
fl->config.root_size, there is a possibility of a buffer overflow.
This is because fl->config.root_size is used while allocating the
buffer and copying the data. If the buffer size during copying is
less than the allocated size, a buffer overflow can occur. Assign
fl->config.root_size to a local variable and use the same variable
for both allocation and copying.

Change-Id: I3b5aadce5ef6c5d9eb1e4f369c6b942d0ff54752
Signed-off-by: Abhishek Singh <quic_abhishes@quicinc.com>