qcacld-3.0: Avoid STA+SAP MCC for non DBS targets
Currently if STA and SAP are operating in SCC and STA moves
to indoor frequency and STA SAP SCC on indoor channel is not
supported then for non DBS targets host fails to stop SAP.

To address this issue, add logic to explicitly invoke
sta_ap_intf_check_work_info work for non DBS targets if STA
is present. This should result in SAP stop if SAP not allowed
to operate in new STA frequency.

CRs-Fixed: 3941081
Change-Id: Ibc7ec922d9032bd31acfc854e647c7d09d678fa0