qcacld-3.0: Abort ongoing/pending scan on SAP start
The rest time argument for scan will be set to zero if no
interface is present, this can lead to keep FW busy in scan
and going off channel to complete the scan. Now if SAP start
is received before scan completes this impacts beaconing of SAP.

Abort scan before SAP start.

Change-Id: I54d6ac9f5461aad16f05c51833005b70a0eb3c02
CRs-Fixed: 3949543