dsp-kernel: Avoid improper address access in DSP
In DSP, va identifies the smaplist between guest OS and userspace.
Using this va, we update the address of the map node in userspace.
If va is passed incorrectly, the update goes wrong, and userspace
tries to access the guest OS address, resulting in a TLB miss.
Change the type of the va parameter from u32 to u64 to ensure va is
properly passed to DSP.

Change-Id: I80993292cf04b60ef89a9a6a228ae3179704a0ad
Signed-off-by: ANANDU KRISHNAN E <quic_anane@quicinc.com>