qcacmn: Make high ap availability as serialization command
Currently high ap availability is not a serialization command
which might result in various unexpected issues.
In current scenario, host driver has given high ap availability
command for LL_LT_SAP to FW and FW is parked on the specific
channel at the same time host driver issued connect command to
FW for the STA interface, since FW can not process this connect
command as it is parked on LL_LT_SAP channel, now when LL_LT_SAP
receives auth/assoc requests from peer, host is not able to process
these requests as it is waiting for the vdev start rsp fo STA
interface. This results in connection delay for STA interface and
also auth/assoc failure for LL_LT_SAP peers.

To address this issue, make high ap availability as serialization
command so that either vdev start goes to the FW or high ap
availability command goes to the FW at a time.

Change-Id: I65025becbec6a30d2eb33d3c23183e84344a7f8a
CRs-Fixed: 3952534