qcacld-3.0: Reject high ap availability if CSA is in progress
Currently there could be scenarios where LL_LT_SAP is
processing CSA and at the same time high ap availability
command is received, host driver and FW accepts this command
but FW will not be able to stay on the current channel and can
move to new channel as part of CSA. Since this request is
accepted by host and FW, EB will try to connect to the LL_LT_SAP
and that will result in connection failure as LL_LT_SAP is
moved to different channel.

To address this issue, add a change to reject high ap availability
command if CSA is in progress.

Change-Id: Iaeda2b8a0c1dabebe999722ad025a7fe5bdfaecb
CRs-Fixed: 3953450