qcacld-3.0: Fix timeout issue during OWE roaming reassociation
During the OWE roaming process, when a reassociation request is
made for the second virtual device (vdev), the command is queued
as “connect” with the subtype “LIM_REASSOC.” However, during error
handling, the LIM_MLM_ASSOC_CNF message is not sent, leading
to a timeout.

This issue needs to be addressed by adding a check for
roaming during the reassociation request and ensuring the
association confirmation is posted.

CRs-Fixed: 3950519
Change-Id: I1e68dee719dd7b8978e11b60e87f69aa9856bb54