qcacld-3.0: flush scan before recovery start
When wlan0 scan is pending on firmware, driver triggers SSR because
of query stats timeout. When driver is doing SSR recovery, the upper
layer set the wlan0 interface down. Driver skips the interface down
because of SSR in progress. And later when driver flush the pending
scan in scan timeout handler, the scan done indication to kernel
is dropped because the IFF_UP flags is not present due to previous
interface down.

Fix by flush scan before driver trigger recovery, so that Pending scan
can be flushed and indicate to kernel.

Change-Id: Ief18c07454ef69eaebcca6a7528aa957708684c9
CRs-Fixed: 3954277