disp: msm: dp: avoid writing DP_DSC_ENABLE if it is already enabled
In DP MST case with DSC enabled, DP re-plug of one of the monitors
can trigger the DSC EN operation to run twice on the other monitor.
When this happens a frame corruption is observed on the other monitor.

This change avoids writing DP_DSC_ENABLE if it is already enabled in
DP MST case.

Change-Id: Ifabf5a413465bd1e5e7fa573969659186e4ddbaa
Signed-off-by: Yahui Wang <quic_yahuiw@quicinc.com>
Signed-off-by: lnxdisplay <lnxdisplay@localhost>