qcacld-3.0: Allow link switch connectivity log for initial connection
For initial connection, the previous link bitmap is not
sent to the Host. This results in rejection of MLO
LINK connectivity log for initial connection which is not
within expectation.

Remove the previous link bitmap check while populating the
parameters for the MLO LINK connectivity log in order to
print MLO LINK connectivity log during initial connection

Change-Id: I0b74d85189b6be825c6f3f20d265112059302962
CRs-Fixed: 3956512