qcacld-3.0: Allow high ap availability command for self peer connection
Currently host driver rejects high ap availability command if there
is any connection in progress. It rejects this command even if
connection is in progress on the same SAP for which this command is
issued. This may lead to various issues.
In current issue, this command is issued by first EB and while
connection from the first EB is in progress second EB sends high
ap availability, host driver rejects this second request as there
is a connection in progress for the first EB because of this, second
EB might fail to connect.
To address this issue, do not reject high AP availability command
if connection is in progress for the same vdev on which this
command is received.

Change-Id: Ic9ed6f3305d298e110d7a609d4489971342d1a14
CRs-Fixed: 3959608