qcacld-3.0: Avoid AP of 4WAY_HANDSHAKE_TIMEOUT
Issue Description:
1. The AP operates on two bands: 2 GHz and 5 GHz.
2. Both bands are configured with WPA2-PSK.
3. The SSID is the same for both bands, but the passwords differ.
4. When a user connects using the 2 GHz band password, the DUT always
selects the 5 GHz band, causing connection failures.

Since the 5 GHz AP has a higher score than the 2 GHz AP while shares same
ssid/AKM, it is always tried first and associated successfully, but fails
during the EAPOL process for REASON_4WAY_HANDSHAKE_TIMEOUT.

Solution: If a deauthentication with REASON_4WAY_HANDSHAKE_TIMEOUT is
received from the AP, add the AP to the avoid list and assign it the
minimum score. This ensures that the framework does not select this
candidate first in subsequent connection attempts.

Change-Id: I580ae3e2f5d057a71c88571f6bb2498e3b31b301
CRs-Fixed: 3958147