qcacmn: Keep 2 GHz in standby when STA connects to 3-link MLO AP
When a STA connects to a 3-link MLO AP, the 6 GHz link
is active, the 2 GHz link becomes inactive, and the 5 GHz
link is in standby. This behavior is unexpected as the
2 GHz link should always be in standby. This commit
checks if the AP is non-DBS and if the 2 GHz link is
not active. If these conditions are met, it sets
the 2 GHz link to standby and the 5 GHz link to inactive.

CRs-Fixed: 3959533
Change-Id: I217fcf3ba1ccd9a160c97f8c3967190f603aaffc