qcacld-3.0: Fix standby link flag in connection info
For sysfs connection info command, standby link flag
retaining the existing value for the current available
link and not getting update with the current link status.

Other issue is logging vdev is NULL in the roaming in
progress check, this is due to standby link don’t have
the valid vdev.

For the first issue, updating the standby link flag for
each link and for the second issue skipping roaming in
progress check for standby link.

Change-Id: I87734247539e521f457de924fcf46dd983841714
CRs-Fixed: 3960149