qcacld-3.0: Set peer as authorize for link reconfig during sta add
If during link reconfig, connection is completed and a disconnect
from NB is received the connection manager state is not connected,
And thus the peer authorize is not sent to firmware.
Leading to no response for vdev stop.

So set peer as authorize for link reconfig during sta add, so that
peer is authorized even if any new NB cmd like disconnect has
changed connection manager SM.

CRs-Fixed: 3960893
Change-Id: I7d80f3998ac41f0b03e083d5c696211f7967e0c4