qcacmn: Mark tx comp ring descriptors invalid during allocation
HW updated the hp without updating the tx descriptor details.
As a result, NULL tx descriptor was obtained causing assert.
To fix this, mark the descriptors with 32 bits invalid value
in higher dword of tx descriptor virtual address which will
be updated by HW if the data has been updated, if not, break
the operation and process in next interrupt when values get
updated.

Change-Id: I69dc3c39d3f18ccd21960c551fbde7ace27b61da
CRs-Fixed: 3980516