qcacmn: Assign wiphy to pdev osif_priv before pdev create
Currently when host driver creates pdev, it invoke regulatory
pdev created notification, from this notification, regulatory
module schedules the registered callbacks. Since wiphy is
assigned to pdev osif priv after pdev create, regulatory
is not scheduing these callbacks as in hdd_dyn_cbk it needs to
access wiphy from pdev osif priv.
During idle restart, host drivers sets reguatory update in progress
as true and this gets reset in hdd_dyn_cbk handling. Since reguatory
is not calling this callback because of the above mentioned reason,
this is not getting reset. Now whenever any operation comes and
check for this flag, it waits for this event to complete which is
never going to happen and eventually leads to timeout.
To address above issue, assign wiphy to pdev osif priv before pdev
create so that when pdev osif priv gets assigned to pdev it has
updated wiphy info and that can be used in hdd_dyn_cbk. Along with
it, invoke the registered reguatory callbacks from regulatory module
so that regulatory update in progress can be reset.

Change-Id: Ieee5907302323ce5c77b2216f1407bf860953e00
CRs-Fixed: 3985656