qcacld-3.0: Fix cpu hotplug scenario for load balancer
The CPU hotplug indication for CPU going down is
received before the CPU goes down. Hence while the
hotplug notifier is being executed, this CPU will be
online.

The CPU hotplug notifier in load balancer generates a
new CPU mask when the last available CPU in its list
is going down. While selecting the new CPUs to be used
for load balancing, the current CPU which is going down
is also being considered, leading to IRQs being affined
to an offline CPU.

To fix the above mentioned issue, ignore the CPU which
goes offline during the selection of CPUs by the load
balancer.

Change-Id: I64fb1262d951a5af4d6a5cb681d748aac71fa146
CRs-Fixed: 3991834