qcacmn: Remove scan_node instead of curr_node
Currently, curr_node is removed when invalid
candidate is removed from candidate list.
As of now it is not causing any issue because
curr_node is the first element of scan_node
entry and address will be same, so free will
not cause any issue as correct address is passed
to free API. But in case if any element is added
before curr_node then address of scan_node and
curr_node will be different which can cause issue
when passed to free due to invalid address.

Fix is to pass scan_node in free API instead of
curr_node to avoid passing invalid address to
free API.

Change-Id: I82f4a5071cf3aa4c2d9af8b26e495ce1f1308141
CRs-Fixed: 3988959