qcacmn: Map GI value of TX TLV to radiotap enum
The GI values in TX TLV do not match the GI values
accepted by the radiotap header due to the presence of
the 0.4us GI field, which is not part of the radiotap standards.

This change maps the 0.8us and 0.4us GI values to the same
radiotap value (0.8us), while 1.6us and 3.2us GI values
are mapped to their respective radiotap values.

Change-Id: Iec751fbfaf88dea3d391b8309aa218fa4a835a13
CRs-Fixed: 3979627