qcacmn: Fix uint8_t data overflow issue
The data type of the num_entries and alloc_size member in the
dp_srng structure is uint32_t, assigning them to type uint8_t
will result in a data overflow error.

Fix is to correct the data type of num_entries and entry_size
to uint32_t.

Additionally, before dividing num_entries, it should be checked
if it is 0.

Change-Id: I3843594426072db8b6afd56f6f6da56853f98e2a
CRs-Fixed: 3995342